Removed circular dependency
