+++ sub-projects/hub/src/server-test/oauth/oauth-router.test.js
@@ -1,7 +1,7 @@
 require('test-globals.js');
 const OAuthTools = require('tools/oauth-tools.js');
 const testUsers = require('tools/users.js');
-const reset = require('utils/db.js').reset;
+const reset = require('utils/uses-db.js').reset;
 
 describe('OAuth2 Router and Integration', () => {
   afterEach(() => reset());
